requirements.txt

matplotlib.pyplot    == 2.5
 tensorflow    == 2.5
 numpy    == 2.5
 random   == 2.5
 glob   == 2.5
 cv2   == 2.5
 os   == 2.5
